+++ src/hooks/useFabricJSEditor.tsx
@@ -102,7 +102,7 @@ const buildEditor = (
       const object = new fabric.Textbox(text, { ...TEXT, fill: strokeColor });
       object.set({ text: text });
       canvas.add(object);
-      // onChange && onChange(JSON.stringify(canvas));
+      onChange && onChange(JSON.stringify(canvas));
     },
     updateText: (text: string) => {
       const objects: any[] = canvas.getActiveObjects();
@@ -125,6 +125,7 @@ const buildEditor = (
         .getActiveObjects()
         .forEach((object) => object.set({ fontSize: textSize }));
       canvas.renderAll();
+      // onChange && onChange(JSON.stringify(canvas));
     },
     changeBoldFont: (apply: boolean) => {
       if (apply) {
@@ -132,6 +133,7 @@ const buildEditor = (
           .getActiveObjects()
           .forEach((object) => object.set({ fontWeight: "bold" }));
         canvas.renderAll();
+        onChange && onChange(JSON.stringify(canvas));
         return;
       }
       canvas
@@ -145,18 +147,21 @@ const buildEditor = (
           .getActiveObjects()
           .forEach((object) => object.set({ fontStyle: "italic" }));
         canvas.renderAll();
+        onChange && onChange(JSON.stringify(canvas));
         return;
       }
       canvas
         .getActiveObjects()
         .forEach((object) => object.set({ fontStyle: "normal" }));
       canvas.renderAll();
+      onChange && onChange(JSON.stringify(canvas));
     },
     changeUnderLineFont: (apply: boolean) => {
       canvas
         .getActiveObjects()
         .forEach((object) => object.set({ underline: apply }));
       canvas.renderAll();
+      // onChange && onChange(JSON.stringify(canvas));
     },
     changeStrikethroughFont: (apply: boolean) => {
       canvas
@@ -168,6 +173,7 @@ const buildEditor = (
       fabric.Image.fromURL(imageUrl, (img) => {
         canvas.add(img);
       });
+      onChange && onChange(JSON.stringify(canvas));
     },
     deleteAll: () => {
       canvas.getObjects().forEach((object) => canvas.remove(object));
